advancedResults middleware
