+++ API_course/controllers/courses.js
@@ -8,24 +8,17 @@ const Bootcamp = require('../models/Bootcamp');
 // @route     GET /api/v1/bootcamps/:bootcampId/courses
 // @access    Public
 exports.getCourses = asyncHandler(async (req, res, next) => {
-    let query;
     if (req.params.bootcampId) {
-        query = Course.find({ bootcamp: req.params.bootcampId })
+        const courses = await Course.find({ bootcamp: req.params.bootcampId });
+
+        return res.status(200).json({
+            success: true,
+            count: courses.length,
+            data: courses
+        });
     } else {
-        // query = await Course.find().populate('bootcamp');//populate to getdata of bootcamp
-        query = await Course.find().populate({
-            path: 'bootcamp',
-            select: 'name description'
-
-        })//populate to getdata of bootcamp
-
+        res.status(200).json(res.advancedResults);
     }
-    const courses = await query;
-    res.status(200).json({
-        success: true,
-        count: courses.length,
-        data: courses
-    });
 });
 // @desc      Get single course
 // @route     GET /api/v1/courses/:id
@@ -76,21 +69,21 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
     let course = await Course.findById(req.params.id);
-  
+
     if (!course) {
-      return next(
-        new ErrorResponse(`No course with the id of ${req.params.id}`),
-        404
-      );
+        return next(
+            new ErrorResponse(`No course with the id of ${req.params.id}`),
+            404
+        );
     }
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true
+        new: true,
+        runValidators: true
     });
-  
+
     res.status(200).json({
-      success: true,
-      data: course
+        success: true,
+        data: course
     });
 });
 
@@ -99,17 +92,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
     const course = await Course.findById(req.params.id);
-  
+
     if (!course) {
-      return next(
-        new ErrorResponse(`No course with the id of ${req.params.id}`),
-        404
-      );
-    } 
+        return next(
+            new ErrorResponse(`No course with the id of ${req.params.id}`),
+            404
+        );
+    }
     await course.remove();
     res.status(200).json({
-      success: true,
-      data: {}
+        success: true,
+        data: {}
     });
-  });
-    
+});
