+++ API_course/routes/courses.js
@@ -9,12 +9,16 @@ const {
 } = require('../controllers/courses');
 
 const Course = require('../models/Course');
+const advancedResults = require('../middleware/advancedResults')
 
 const router = express.Router({ mergeParams: true });//to add => /bootcamps/:id/courses/
 
 router
     .route('/')
-    .get(getCourses)
+    .get(advancedResults(Course, {
+        path: 'bootcamp',
+        select: 'name description'
+    }), getCourses)
     .post(addCourse)
 router
     .route('/:id')
